refactor: room entity로 전환
