+++ src/main/java/chess/web/controller/dto/RoomResponseDto.java
@@ -1,17 +1,24 @@
 package chess.web.controller.dto;
 
-import chess.domain.Room;
+import chess.domain.entity.Room;
 
 public class RoomResponseDto {
-    public static class Title{
+    public static class TitleAndId{
         private final String title;
+        private final Long id;
 
-        public Title(Room room){
+        public TitleAndId(Room room) {
             this.title = room.getTitle();
+            this.id = room.getId();
         }
 
         public String getTitle() {
             return title;
         }
+
+        public Long getId() {
+            return id;
+        }
+
     }
 }
